supabowl/Downloads/Cursor/budget_allocation_mcp/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
pydantic[email]>=1.8.2  # Added for achievement system
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
requests>=2.26.0
aiohttp>=3.8.1
plaid-python>=7.2.1
intuit-oauth>=1.2.3
python-jose>=3.3.0  # For JWT handling
passlib>=1.7.4      # For password hashing
python-multipart>=0.0.5  # For form data handling
APScheduler>=3.8.1  # Added for achievement system
redis>=4.0.2        # Added for achievement system
pytest>=6.2.5       # For testing
pytest-asyncio>=0.15.1  # For async testing
black>=21.7b0       # For code formatting
isort>=5.9.3        # For import sorting
flake8>=3.9.2       # For linting